service2_processor/requirements.txt
requests
python-dotenv
supabase
flask